libft update

+++ libft/ft_strdup.c
@@ -19,8 +19,8 @@ char	*ft_strndup(const char *s1, int len)
 
 	if (!s1)
 		return (NULL);
-	if (len > ft_strlen(s1))
-		len = ft_strlen(s1);
+	if (len > (int)ft_strlen(s1))
+		len = (int)ft_strlen(s1);
 	i = 0;
 	copy = (char *)malloc(sizeof(char) * ft_strlen((s1) + 1));
 	if (!copy)
